updator/requirements.txt
PyYAML==6.*
requests==2.*
packaging==21.*
